tools/agstream_manager/pyproject.toml
[project]
name = "agstream-manager"
version = "0.1.0"
description = "AGStream Manager - Web UI for managing Kafka streams, schemas, and Flink SQL"
requires-python = ">=3.11,<3.13"
dependencies = [
    "agentics-py @ {root:uri}/../..",
    "flask>=3.0.0",
    "flask-cors>=5.0.0",
    "flask-sock>=0.7.0",
    "streamlit>=1.45.1",
    "psutil>=6.1.0",
    "requests>=2.32.0",
    "pydantic>=2.0.0",
    "kafka-python-ng>=2.2.0",
    "python-dotenv>=1.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.0",
    "pytest-asyncio>=1.0.0",
    "pytest-html>=4.1.1",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--html=report.html --self-contained-html"
markers = [
    "agstream: AGStream integration tests (requires Kafka/Schema Registry)",
    "flink: Tests requiring Flink Docker environment (pyflink, semantic_operators)",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.uv]
package = false
